label value validation
