+++ pkg/aws/common_mutator_test.go
@@ -6,171 +6,12 @@ import (
 	"strconv"
 	"testing"
 
-	releasev1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/release/v1alpha1"
 	"github.com/giantswarm/micrologger/microloggertest"
 
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/label"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/mutator"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/unittest"
 )
-
-func Test_MaxBatchSizeIsValid(t *testing.T) {
-	testCases := []struct {
-		name  string
-		input string
-		valid bool
-	}{
-		{
-			name:  "case 0: int - simple value",
-			input: "5",
-			valid: true,
-		},
-		{
-			name:  "case 1: int - big value",
-			input: "200",
-			valid: true,
-		},
-		{
-			name:  "case 2: int - invalid value - negative number",
-			input: "-10",
-			valid: false,
-		},
-		{
-			name:  "case 2: int - invalid value - zero",
-			input: "0",
-			valid: false,
-		},
-		{
-			name:  "case 4: percentage - simple value",
-			input: "0.5",
-			valid: true,
-		},
-		{
-			name:  "case 5: percentage - rounding",
-			input: "0.35",
-			valid: true,
-		},
-		{
-			name:  "case 6: percentage - rounding",
-			input: "0.32",
-			valid: true,
-		},
-		{
-			name:  "case 7: percentage - invalid value - too big",
-			input: "1.5",
-			valid: false,
-		},
-		{
-			name:  "case 8: percentage - invalid value - negative",
-			input: "-0.5",
-			valid: false,
-		},
-		{
-			name:  "case 9: invalid value - '50%'",
-			input: "50%",
-			valid: false,
-		},
-		{
-			name:  "case 10: invalid value - string",
-			input: "test",
-			valid: false,
-		},
-		{
-			name:  "case 11: invalid value - number and string",
-			input: "5erft",
-			valid: false,
-		},
-		{
-			name:  "case 12: invalid value - float and string",
-			input: "0.5erft",
-			valid: false,
-		},
-	}
-
-	for i, tc := range testCases {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			isValid := MaxBatchSizeIsValid(tc.input)
-
-			if isValid != tc.valid {
-				t.Fatalf("%s - expected '%t' got '%t'\n", tc.name, tc.valid, isValid)
-			}
-		})
-	}
-}
-
-func Test_PauseTimeIsValid(t *testing.T) {
-	testCases := []struct {
-		name  string
-		value string
-		valid bool
-	}{
-		{
-			name:  "case 0: simple value",
-			value: "PT15M",
-			valid: true,
-		},
-		{
-			name:  "case 2: simple value",
-			value: "PT10S",
-			valid: true,
-		},
-		{
-			name:  "case 3: simple value",
-			value: "PT2M10S",
-			valid: true,
-		},
-		{
-			name:  "case 4: simple value",
-			value: "PT2M10S",
-			valid: true,
-		},
-		{
-			name:  "case 5: invalid value value",
-			value: "10m",
-			valid: false,
-		},
-		{
-			name:  "case 6: invalid value value",
-			value: "10s",
-			valid: false,
-		},
-		{
-			name:  "case 7: invalid value value",
-			value: "10",
-			valid: false,
-		},
-		{
-			name:  "case 8: invalid value value",
-			value: "1 hour",
-			valid: false,
-		},
-		{
-			name:  "case 9: invalid value value",
-			value: "random string",
-			valid: false,
-		},
-		{
-			name:  "case 10: invalid value value",
-			value: "",
-			valid: false,
-		},
-		{
-			name:  "case 11: duration too big",
-			value: "PT1H2M",
-			valid: false,
-		},
-	}
-
-	for i, tc := range testCases {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			result := PauseTimeIsValid(tc.value)
-
-			if result != tc.valid {
-				t.Fatalf("%s -  expected '%t' got '%t'\n", tc.name, tc.valid, result)
-			}
-		})
-	}
-}
 
 func TestLabelFromCluster(t *testing.T) {
 	testCases := []struct {
@@ -341,106 +182,6 @@ func TestLabelFromRelease(t *testing.T) {
 			// check if the release label is as expected
 			if tc.expectedPatch != updatedOperator {
 				t.Fatalf("expected %#q to be equal to %#q", tc.expectedPatch, updatedOperator)
-			}
-		})
-	}
-}
-
-func TestNewestReleaseVersion(t *testing.T) {
-	testCases := []struct {
-		ctx  context.Context
-		name string
-
-		releases        []unittest.ReleaseData
-		expectedVersion string
-	}{
-		{
-			// take only active releases
-			name: "case 0",
-			ctx:  context.Background(),
-
-			releases: []unittest.ReleaseData{
-				{
-					Name:  "v1.2.3",
-					State: releasev1alpha1.StateActive,
-				},
-				{
-					Name:  "v2.1.0",
-					State: releasev1alpha1.StateDeprecated,
-				},
-			},
-			expectedVersion: "1.2.3",
-		},
-		{
-			// don't take dev versions
-			name: "case 1",
-			ctx:  context.Background(),
-
-			releases: []unittest.ReleaseData{
-				{
-					Name:  "v1.2.3",
-					State: releasev1alpha1.StateActive,
-				},
-				{
-					Name:  "v3.2.3-dev",
-					State: releasev1alpha1.StateActive,
-				},
-			},
-			expectedVersion: "1.2.3",
-		},
-		{
-			// sort releases correctly
-			name: "case 1",
-			ctx:  context.Background(),
-
-			releases: []unittest.ReleaseData{
-				{
-					Name:  "v1.1.1",
-					State: releasev1alpha1.StateActive,
-				},
-				{
-					Name:  "v0.2.3",
-					State: releasev1alpha1.StateActive,
-				},
-				{
-					Name:  "v1.2.3",
-					State: releasev1alpha1.StateActive,
-				},
-				{
-					Name:  "v1.2.1",
-					State: releasev1alpha1.StateActive,
-				},
-			},
-			expectedVersion: "1.2.3",
-		},
-	}
-	for i, tc := range testCases {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			var err error
-
-			fakeK8sClient := unittest.FakeK8sClient()
-			handle := &Handler{
-				K8sClient: fakeK8sClient,
-				Logger:    microloggertest.New(),
-			}
-			// create releases for testing
-			for _, r := range tc.releases {
-				release := unittest.DefaultRelease()
-				release.SetName(r.Name)
-				release.Spec.State = r.State
-				err = fakeK8sClient.CtrlClient().Create(tc.ctx, &release)
-				if err != nil {
-					t.Fatal(err)
-				}
-			}
-			// run fetcher to get newest active release version
-			version, err := FetchNewestReleaseVersion(handle)
-			if err != nil {
-				t.Fatal(err)
-			}
-			// check if the release version is as expected
-			if tc.expectedVersion != version.String() {
-				t.Fatalf("expected %#q to be equal to %#q", tc.expectedVersion, version.String())
 			}
 		})
 	}
